Refactored JNDI code to support all backends.
In order to keep it independent of any particular application server I had to remove the Glassfish-specific code. I want to "undelete" the associated documentation as a Glassfish-specific recipe, but first I need to actually understand it.

Also added a simple tomcat management command which by now only helps with the JNDI configuration but should grow more features on the near future (like deployment, reloading, etc.). It also served to drive the refactoring of the common zxJDBC connection code to the new zxJDBCDatabaseWrapper base class.

git-svn-id: http://django-jython.googlecode.com/svn/branches/jndi@163 fc56e334-f652-0410-822d-0b1548283ca8